Add hyper test
